Refactor frontend to more traditional Flask organization.
