+++ mergebot_frontend/static/css/mergebot.css
