+++ mergebot_frontend/templates/item.html
@@ -1,0 +1,26 @@
+{% extends "base.html" %}
+{% block title %}Apache Mergebot: {{ poller.project_name }}{% endblock %}
+{% block content %}
+<div class="container">
+    <div class="row">
+        <div class="col-xs-12">
+            <h1>Merge Status For PR# {{ item_statuses[0].item_id }}</h1>
+        </div>
+    </div>
+    <div class="row">
+        {% for item in item_statuses %}
+        <div class="col-xs-12">
+            <div class="panel panel-default">
+                <div class="panel-heading">
+                    <h3 class="panel-title pull-left">Status: {{ item.status }}</h3>
+                    <div class="clearfix"></div>
+                </div>
+                <div class="panel-body">
+                    Updated: {{ item.timestamp }}
+                </div>
+            </div>
+        </div>
+        {% endfor %}
+    </div>
+</div>
+{% endblock %}
