+++ run_frontend.py
@@ -1,0 +1,4 @@
+from mergebot_frontend import app
+
+if __name__ == "__main__":
+    app.run(debug=True, host='0.0.0.0', port='8080')
